pyproject.toml
[build-system]
requires = ["setuptools==70.3.0"]
build-backend = "setuptools.build_meta"

[project]
name = "taina"
dynamic = ["version"]
dependencies = [
    "SQLAlchemy==2.0.32",
    "alembic==1.13.2",
    "fastapi==0.112.0",
    "psycopg==3.1.20",
    "psycopg_pool==3.2.2",
    "pydantic-settings==2.4.0",
    "pydantic==2.8.2",
    "pyjwt==2.9.0",
    "python-multipart==0.0.9",
    "uvicorn==0.30.5",
    "uvloop==0.19.0",
]

[project.optional-dependencies]
dev = [
    "async-asgi-testclient==1.4.11",
    "pytest-asyncio==0.23.8",
    "pytest-cov==5.0.0",
    "pytest-freezegun==1.5.1",
    "pytest-mock==3.14.0",
    "pytest-xdist==3.6.1",
    "pytest==8.3.2",
    "ruff==0.5.7",
]

[tool.setuptools.dynamic]
version = {attr = "taina.__version__"}

[tool.coverage.run]
omit = ["taina/migrations/*"]

[tool.pytest.ini_options]
addopts = "-n 4 --cov=taina --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    'ignore:distutils Version classes are deprecated:DeprecationWarning',
]

[tool.ruff.lint]
select = [
    "F", "E", "W", "C90", "I", "N", "B", "A", "Q", "RET", "ERA", "RUF",
    "COM", "C4", "DTZ", "G", "PIE", "T20", "PT", "RSE", "SIM", "W",
]
ignore = ["N818", "B904", "B008"]

[tool.ruff.lint.isort]
force-single-line = true
